requirements.txt
PyGithub
openapi-spec-validator
pyyaml
prance